+++ frontend/src/features/auth/index.ts
@@ -1,6 +1,6 @@
 // Components
 export { UserFormDialog } from './components/UserFormDialog';
-export { UsersManagement } from './components/UsersManagement';
+export { UsersManagement } from './components/UsersTab';
 
 // Hooks
 export { useUsers } from './hooks/useUsers';
